@@ -6,97 +6,73 @@
     "start": "ng serve --port 3000 --ssl",
     "start:insecure": "ng serve --port 3000",
     "build": "ng build",
-    "build:prod": "ng build --configuration production --base-href \"/metaforms/\"",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "build:prod": "ng build --configuration production --base-href \"/metaforms/\""
   },
   "private": true,
-  "engines": {
-    "node": ">=14 <15",
-    "npm": "6.14.17"
-  },
   "dependencies": {
-    "@angular/animations": "~12.2.16",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "~12.2.16",
-    "@angular/compiler": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.16",
-    "@angular/localize": "~12.2.16",
-    "@angular/material": "^12.2.13",
-    "@angular/material-moment-adapter": "^12.2.13",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/router": "~12.2.16",
-    "@cinchy-co/angular-sdk": "6.0.0-alpha.0",
-    "@fortawesome/angular-fontawesome": "~0.9.0",
+    "@angular/animations": "^18.2.11",
+    "@angular/cdk": "^18.2.12",
+    "@angular/common": "^18.2.11",
+    "@angular/compiler": "^18.2.11",
+    "@angular/core": "^18.2.11",
+    "@angular/forms": "^18.2.11",
+    "@angular/localize": "^18.2.11",
+    "@angular/material": "^18.2.12",
+    "@angular/material-moment-adapter": "^18.2.12",
+    "@angular/platform-browser": "^18.2.11",
+    "@angular/platform-browser-dynamic": "^18.2.11",
+    "@angular/router": "^18.2.11",
+    "@cinchy-co/angular-sdk": "^7.0.0-rc.0",
+    "@ctrl/ngx-codemirror": "^7.0.0",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-regular-svg-icons": "^6.2.0",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
-    "@tiptap/core": "2.0.0-beta.212",
-    "@tiptap/extension-bullet-list": "2.0.3",
-    "@tiptap/extension-code-block-lowlight": "2.0.0-beta.212",
-    "@tiptap/extension-image": "2.0.0-beta.212",
-    "@tiptap/extension-link": "2.0.0-beta.212",
-    "@tiptap/extension-ordered-list": "2.0.3",
-    "@tiptap/extension-table": "2.0.0-beta.212",
-    "@tiptap/extension-table-cell": "2.0.0-beta.212",
-    "@tiptap/extension-table-header": "2.0.0-beta.212",
-    "@tiptap/extension-table-row": "2.0.0-beta.212",
-    "@tiptap/extension-task-item": "2.0.0-beta.212",
-    "@tiptap/extension-task-list": "2.0.0-beta.212",
-    "@tiptap/extension-text": "2.0.3",
-    "@tiptap/extension-underline": "2.0.0-beta.212",
-    "@tiptap/pm": "2.0.0-beta.212",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.1",
+    "@popperjs/core": "^2.11.8",
+    "@tiptap/core": "2.9.1",
+    "@tiptap/extension-code-block-lowlight": "2.9.1",
+    "@tiptap/extension-image": "2.9.1",
+    "@tiptap/extension-link": "2.9.1",
+    "@tiptap/extension-table": "2.9.1",
+    "@tiptap/extension-table-cell": "2.9.1",
+    "@tiptap/extension-table-header": "2.9.1",
+    "@tiptap/extension-table-row": "2.9.1",
+    "@tiptap/extension-task-item": "2.9.1",
+    "@tiptap/extension-task-list": "2.9.1",
+    "@tiptap/extension-underline": "2.9.1",
+    "@tiptap/pm": "2.9.1",
     "@tiptap/prosemirror-tables": "^1.1.4",
-    "@tiptap/starter-kit": "2.0.0-beta.212",
+    "@tiptap/starter-kit": "2.9.1",
     "@uiowa/digit-only": "^3.2.1",
-    "angular2-multiselect-dropdown": "^4.6.5",
-    "brace": "^0.11.1",
+    "codemirror": "^5.65.16",
     "file-saver": "^2.0.2",
     "font-awesome": "^4.7.0",
     "lodash": "^4.17.21",
     "lowlight": "^2.8.0",
-    "moment": "^2.29.4",
-    "ng-pick-datetime": "^7.0.0",
-    "ng2-ace-editor": "^0.3.9",
-    "ngx-filter-pipe": "^2.1.2",
-    "ngx-mat-select-search": "^4.2.1",
+    "moment": "^2.30.1",
+    "ngx-mat-select-search": "^7.0.6",
     "ngx-numeral": "^1.0.0",
-    "ngx-spinner": "^12.0.0",
+    "ngx-spinner": "^16.0.2",
+    "pdfmake": "^0.2.15",
+    "ramda": "^0.28.0",
+    "rxjs": "^6.6.7",
     "ngx-toastr": "^14.3.0",
-    "pdfmake": "^0.1.70",
-    "prosemirror-model": "1.19.2",
-    "prosemirror-transform": "1.7.3",
-    "prosemirror-view": "1.31.5",
-    "ramda": "^0.27.2",
-    "rxjs": "~6.5.4",
+    "prosemirror-model": "1.23.0",
+    "prosemirror-transform": "1.10.2",
+    "prosemirror-view": "1.36.0",
     "tslib": "^2.0.0",
     "util": "^0.12.4",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.18",
-    "@angular/cli": "~12.2.18",
-    "@angular/compiler-cli": "~12.2.16",
-    "@angular/language-service": "~12.2.16",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
+    "@angular/build": "^18.2.11",
+    "@angular/cli": "^18.2.11",
+    "@angular/compiler-cli": "^18.2.11",
+    "@angular/language-service": "^18.2.11",
+    "@types/node": "^18.19.1",
     "file-loader": "^6.2.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
   }
 }
