@@ -14,7 +14,7 @@ export class ErrorService {
     return error?.cinchyException?.data?.details ?? error?.cinchyException?.message ?? error.message ?? "";
 
     // In the future, we will want to include both pieces of information as part of the same notification, but for now
-    // we just want a simplified breakdown which prioritizes the most relevatn information
+    // we just want a simplified breakdown which prioritizes the most relevant information
     /*
      * const message = error?.cinchyException?.data?.details || null;
      * const details = error?.cinchyException?.message || error?.message || null;
