@@ -1,36 +1,35 @@
 import { NgModule } from "@angular/core";
+import { OverlayModule } from "@angular/cdk/overlay";
 import { ScrollingModule } from "@angular/cdk/scrolling";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { BrowserModule } from "@angular/platform-browser";
 import { RouterModule } from "@angular/router";
 
-import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
-
 import { NgxMatSelectSearchModule } from "ngx-mat-select-search";
+
+import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 
 import { CustomMaterialModule } from "../custom-material.module";
 
 import { SearchDropdownComponent } from "./search-dropdown/search-dropdown.component";
-import { TypeaheadComponent } from "./typeahead/typeahead.component";
 
 
 @NgModule({
   declarations: [
-    SearchDropdownComponent,
-    TypeaheadComponent
+    SearchDropdownComponent
   ],
   imports: [
     BrowserModule, // Used for async pipes
     CustomMaterialModule,
-    NgxMatSelectSearchModule,
     FontAwesomeModule,
     FormsModule,
+    OverlayModule,
     ReactiveFormsModule,
     RouterModule,
-    ScrollingModule
+    ScrollingModule,
+    NgxMatSelectSearchModule
   ],
   exports: [
-    TypeaheadComponent,
     SearchDropdownComponent
   ],
   providers: []
