@@ -1,8 +1,14 @@
-import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { enableProdMode } from "@angular/core";
+import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 
-import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
+import { AppModule } from "./app/app.module";
+import { environment } from "./environments/environment";
+
+
+import "codemirror/mode/javascript/javascript";
+import "codemirror/mode/sql/sql";
+import "codemirror/mode/xml/xml";
+
 
 if (environment.production) {
   enableProdMode();
